change choice actions
