+++ assets/src/components/activities/multiple_choice/actions.ts
@@ -7,11 +7,12 @@ import {
   createMatchRule,
   getResponses,
 } from 'components/activities/common/responses/authoring/responseUtils';
+import { getChoice, getChoices } from 'components/activities/common/choices/authoring/choiceUtils';
 
 export const MCActions = {
   addChoice(choice: Choice) {
     return (model: MultipleChoiceModelSchema, post: PostUndoable) => {
-      ChoiceActions.addChoice(choice)(model, post);
+      ChoiceActions.addChoice(choice)(model);
 
       model.authoring.parts[0].responses.push(makeResponse(`input like {${choice.id}}`, 0, ''));
     };
@@ -19,11 +20,25 @@ export const MCActions = {
 
   removeChoice(id: string) {
     return (model: MultipleChoiceModelSchema, post: PostUndoable) => {
-      ChoiceActions.removeChoice(id)(model, post);
+      const choice = getChoice(model, id);
+      const index = getChoices(model).findIndex((c) => c.id === id);
+      ChoiceActions.removeChoice(id)(model);
 
       model.authoring.parts[0].responses = getResponses(model).filter(
         (r) => r.rule !== createMatchRule(id),
       );
+
+      post({
+        description: 'Removed a choice',
+        operations: [
+          {
+            path: '$.choices',
+            index,
+            item: JSON.parse(JSON.stringify(choice)),
+          },
+        ],
+        type: 'Undoable',
+      });
     };
   },
